refactor(timeline): colocate the new hooks with their components
useTimelineViewportSize and useScrollEdgeFade are only used by
TimelineOverview and TimelineToolbar respectively - move them out of
the general src/hooks/ folder and into the same directory as their
components.
