+++ src/pages/EditionView/tabs/ScheduleTab/horizontal/useScrollEdgeFade.ts
