+++ src/pages/EditionView/tabs/ScheduleTab/horizontal/useTimelineViewportSize.ts
